Simplify zenoh shutdown logic (#947)
