+++ include/gz/transport/RepHandler.hh
@@ -59,6 +59,7 @@ namespace gz::transport
   /// \class IRepHandler RepHandler.hh gz/transport/RepHandler.hh
   /// \brief Interface class used to manage a replier handler.
   class GZ_TRANSPORT_VISIBLE IRepHandler
+    : public std::enable_shared_from_this<IRepHandler>
   {
     /// \brief Constructor.
     /// \param[in] _pUuid Process UUID.
@@ -98,15 +99,6 @@ namespace gz::transport
     public: virtual std::string RepTypeName() const = 0;
 
 #ifdef HAVE_ZENOH
-    /// \brief Temporarily store the dispatch closure. Must be called by the
-    /// most derived RepHandler subclass before CreateZenohQueriable.
-    /// The dispatch closure should capture a weak_ptr to the handler to safely
-    /// abort if invoked during handler destruction.
-    protected: void SetZenohQueryableDispatch(
-      const std::string &_service,
-      std::function<bool(const std::string &request,
-                         std::string &response)> _dispatch);
-
     /// \brief Create a Zenoh queriable.
     /// \param[in] _session Zenoh session.
     /// \param[in] _service The service.
@@ -135,8 +127,7 @@ namespace gz::transport
   // the service call. 'Rep' is the protobuf message type that will be filled
   /// with the service response.
   template <typename Req, typename Rep> class RepHandler
-    : public IRepHandler,
-      public std::enable_shared_from_this<RepHandler<Req, Rep>>
+    : public IRepHandler
   {
     // Documentation inherited.
     using IRepHandler::IRepHandler;
@@ -168,15 +159,6 @@ namespace gz::transport
       const std::string &_service)
     {
       this->SetCallback(std::move(_cb));
-      std::weak_ptr<RepHandler<Req, Rep>> weakSelf = this->weak_from_this();
-      this->SetZenohQueryableDispatch(_service,
-        [weakSelf](const std::string &request, std::string &response) -> bool
-        {
-          auto self = weakSelf.lock();
-          if (!self)
-            return false;
-          return self->RunCallback(request, response);
-        });
       this->CreateZenohQueriable(_session, _service);
     }
 #endif
